Refactored scope resolution and refined the path grammar
Introduced 'lazy scopes' that resolve their contents partially as requested. The previous approach had a number of shortcomings w.r.t. redeclaration checking (conflicting declarations were not reported immediately) and forward references (the previous design forced most the data of declaration nodes to be deferred until the full scope was available).

The current design resolves declarations as needed (see LazyScope.resolve). The resolution code now has access to the (partially resolved) parent scope. Conflicting declarations within the same scope are handled generically in LazyScope. Function overloading is not yet supported and two functions with the same name will show up as a conflicting declaration.
Conflicting declarations in two different scopes (e.g. shadowing a variable in a derived component) have more complex rules and are not yet implemented.

The code for bottom-up resolution (building the declarations starting from an arbitrary PSI node and recursively going back up in the PSI tree) has been removed for now, as it was not needed. It should be relatively simple to add.

Refined the path grammar to produce a right-associative hierarchy of sub-paths that is more suited to the path resolution process.