Refactored {Type,Declaration}Resolver
The responsibility of the type resolver is to check and annotate expression nodes with correct types. The declaration resolver creates the *Declaration nodes.
The class ReferenceResolver now contains all methods related to name resolution in scopes.
The goal is simply to put related semantic code together (Type checking, declaration resolution, name resolution), but sometimes it's not clear where a method should go. (e.g. now the TypeResolver also creates LocalVariableDeclarations to avoid a cyclic dependency on the DeclarationResolver)
Another solution would be to put semantic checking code in the Declaration nodes.