Two-pass semantic verifications
The sementic pass has been split into two explicit sub-passes: the declaration pass (implemented in DeclarationResolver), which resolves the declarations and some types when they are explicitly specified, and the body pass (BodyResolver) which resolves the initializer expressions and function bodies.
When a type cannot be computed in the declaration pass, an instance of DeferredType is created instead, which contains the procedure for computing the type after the declaration pass.